dataperf/dataperf_acquisition/requirements.txt
fastapi==0.63.0
mangum==0.10.0
uvicorn==0.13.2
websockets==8.1
scikit-learn==1.1.2
numpy==1.23.2
boto3==1.24.75
pandas==1.5.0
python-dotenv==0.21.0
pyarrow==9.0.0
tqdm==4.64.1
PyYAML==5.4.1
